Fix economy balance casts and AddAmmoPacks flow
Cast economy balances to int when reading GetPlayerBalance to avoid type/mismatch issues and ensure _balanceCache stores an int. Restructure AddAmmoPacks: normalize braces/formatting, resolve API usage to pass IPlayer instead of playerId for AddPlayerBalance/SaveData, cast returned balance to int, wrap economy calls in try/catch with improved logging, and move the "bridge offline" fallback to the else branch so players are notified when the economy bridge is unavailable.

+++ src/ZombiePlagueLegacyCS2/ZPL.AmmoPacks.Service.cs
@@ -40,7 +40,7 @@ public class AmmoPacksService
 
         try
         {
-            _balanceCache[id] = Math.Max(0, _api.GetPlayerBalance(player, WalletKind));
+            _balanceCache[id] = Math.Max(0, (int)_api.GetPlayerBalance(player, WalletKind));
         }
         catch { _balanceCache[id] = 0; }
     }
@@ -61,7 +61,7 @@ public class AmmoPacksService
 
         try
         {
-            int balance = Math.Max(0, _api.GetPlayerBalance(player, WalletKind));
+            int balance = Math.Max(0, (int)_api.GetPlayerBalance(player, WalletKind));
             _balanceCache[playerId] = balance;
             return balance;
         }
